[ISSUE #9588]♻️Consolidate batch delete protocol tests
